requirements.txt
requests==2.32.3
python-dotenv==1.0.1
urllib3==1.26.6
python-telegram-bot==13.0